requirements.txt
# Requirements for the Privacy Camera application
# Single command:
#   pip install -r requirements.txt

PyQt6>=6.6.0
numpy>=1.24.0
opencv-python>=4.8.0

torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Use GPU-enabled ONNX Runtime (also works on CPU, but requires CUDA libs present)
onnxruntime-gpu>=1.15.0
